+++ src/sim/render/camera/sensor.rs
@@ -8,19 +8,19 @@ use std::fmt::{Display, Formatter, Result};
 #[derive(Debug)]
 pub struct Sensor {
     /// Image resolution.
-    res: (usize, usize),
+    res: (u64, u64),
     /// Optional sub-sampling power.
     super_sample_power: Option<i32>,
 }
 
 impl Sensor {
-    clone!(res, (usize, usize));
+    clone!(res, (u64, u64));
     clone!(super_sample_power, Option<i32>);
 
     /// Construct a new instance.
     #[inline]
     #[must_use]
-    pub fn new(aspect_ratio: &AspectRatio, hr_res: usize, super_sample_power: Option<i32>) -> Self {
+    pub fn new(aspect_ratio: &AspectRatio, hr_res: u64, super_sample_power: Option<i32>) -> Self {
         debug_assert!(super_sample_power.is_none() || super_sample_power.unwrap() > 1);
 
         Self {
@@ -32,7 +32,7 @@ impl Sensor {
     /// Calculate the total number of pixels.
     #[inline]
     #[must_use]
-    pub const fn num_pixels(&self) -> usize {
+    pub const fn num_pixels(&self) -> u64 {
         self.res.0 * self.res.1
     }
